Add parameterization to @httpMalformedRequestTests
As an opt-in feature of @httpMalformedRequestTests, test authors can use the
testParameters field to create parameterized tests, for testing multiple
instances of the same kind of bad input, such as multiple forms of invalid
booleans. The expansion of these test cases is performed within the trait,
and code generators are not exposed to the parameterized test cases at any
point.